refactor: move pastedImageCache and paste dispatch onto the controller
The paste-image path now delegates to launchImagePaste, mirroring the
bottom-sheet dispatch. pastedImageCache and its savedInstanceState
plumbing move with it; the fragment forwards save/restore through
onSaveInstanceState / onRestoreInstanceState on the controller.
